Alpha v2.0.0 a3 Initial plugin to create the elemental resources for livestream or VOD
- Started to decouple duplicate functions.

Please note this uses lambda functions to deploy the resources as there is no cloudformation support for elemental yet.
